feat(Dependency): complete rework of system
The old dependency system was trapped in a place of confusing and
inconsistent language, making it unmaintainable going forward by anyone
other than myself (the original author), also making it impossible for
end users to implement custom dependency resolution rules or logic.

Goals:

1:
Detangle the resolution of different types of dependency from the
resolution processing logic.

This was the major pain point, making the system both unmaintainable
and inextensible, as it meant that for any new custom logic, internal
files needed to be modified as well, in a manner that was not easy to
follow, and made the core resolution logic hard to read.

2:
Improve language consistency across resolution strategies.

Inconsistent and unhelpful names made it difficult to figure out which
dependency implementation to reach for without going to look at another
example.

3:
Enable better logical groupings of dependencies.

Previously all dependencies were grouped as `and`, now both `and` and
`or` are supported, with the ability to easily add more boolean logic
comparisons to the system. While the older iteration was good for the
majority of situations, it was likely to seriously damage flexibility
for complex resolution systems.

4:
Improve resolution callback capability.

Resolution callback functionality was inconsistently available, was
added through a large amount of painful handwritten boilerplate to the
builder process, and was not able to be added to the result of things
like the new short circuiting `or` operator.

Areas for improvement:

Arguments could easily be made in favour of adding an additional
argument or two to the resolution function signature, this should be
reviewed now, before the library is published.

current arguments:
- OpModeWrapper
- currently resolved Features
- if the processor is in a yielding cycle

potential additional arguments:
- unresolved Features still in the processor

Additionally, KDoc coverage could be improved.