feat(TestFixtures): Now publishes TestFixtures
The TestFixtures are organised to enable a Dairy OpMode Runtime for
tests, the TestOpMode abstract class can be used. This allows unit tests
to run that rely on Features. This also allows Features to be used to
mock additional test requirements (e.g. the hardwaremap), or to test the
state of other Features, or the OpMode / User Code.

This runtime does not mock the hardwaremap or telemetry objects right
now.

This runtime does not currently support wrapping an actual non-test
OpMode in order to run tests against a non-test OpMode.

Further support may include:
- mocking telemetry.
- easy support for mocking the hardwaremap.
- easy support for mocking gamepads.
- easy support for wrapping and testing a non-test OpMode

Many of these features will be easier for many Dairy eco-system
libraries to emulate.

Unit tests do not support Sinister, which means any sinister detections
must be done by hand. At the moment, TestOpMode must be told which
features exist, and other Sinister behaviours must be hand-recreated or
mocked themselves.