fix: DependencyBase, version bump to 1.1.1
Removed the need for DependencyBase usage in general. It still comes in
useful for delegation and composition.

Dependencies now have the capabilities to perform the same operations as
DependencyBase added, and now resolution and failure callbacks are
performed using function composition, rather than mutable lists of
callbacks.

This also fixes issues to do with potential errors thrown in callback
resolution, and ensure that the correct callbacks are called.